Reformat code; Optimize tests

+++ HexCore/HexGraph/Graph.cs
@@ -10,16 +10,6 @@ namespace HexCore.HexGraph
     [Serializable]
     public class Graph : IWeightedGraph
     {
-        private readonly OffsetTypes _offsetType;
-        public readonly List<List<CellState>> Columns = new List<List<CellState>>();
-        private int _width;
-        private int _height;
-        private readonly List<MovementType> _movementTypes;
-
-        private List<Coordinate3D> _blocked;
-
-        private List<Coordinate3D> _cubeCoordinates;
-
         // Possible directions to detect neighbors        
         private readonly List<Coordinate3D> _directions = new List<Coordinate3D>
         {
@@ -28,14 +18,35 @@ namespace HexCore.HexGraph
             new Coordinate3D(0, +1, -1),
             new Coordinate3D(-1, +1, 0),
             new Coordinate3D(-1, 0, +1),
-            new Coordinate3D(0, -1, +1),
+            new Coordinate3D(0, -1, +1)
         };
+
+        private readonly List<MovementType> _movementTypes;
+        private readonly OffsetTypes _offsetType;
+        public readonly List<List<CellState>> Columns = new List<List<CellState>>();
+
+        private List<Coordinate3D> _blocked;
+
+        private List<Coordinate3D> _cubeCoordinates;
+        private int _height;
+        private int _width;
 
         public Graph(int width, int height, OffsetTypes offsetType, List<MovementType> movementTypes)
         {
             _offsetType = offsetType;
             _movementTypes = movementTypes;
             Resize(width, height);
+        }
+
+        public IEnumerable<Coordinate3D> GetPassableNeighbors(Coordinate3D position)
+        {
+            return GetNeighbors(position, true);
+        }
+
+        public int GetMovementCost(Coordinate3D coordinate, MovementType unitMovementType)
+        {
+            var cellState = GetCellStateByCoordinate3D(coordinate);
+            return unitMovementType.GetCostTo(cellState.MovementType.Name);
         }
 
         private void ResizeColumn(int columnNumber, int newSize)
@@ -69,57 +80,48 @@ namespace HexCore.HexGraph
         public void Resize(int width, int height)
         {
             if (width > _width)
-            {
                 for (var columnNumber = _width; columnNumber < width; columnNumber++)
-                {
                     AddNewColumn(height);
-                }
-            }
 
-            if (_width > width)
-            {
-                Columns.RemoveRange(width, _width - width);
-            }
+            if (_width > width) Columns.RemoveRange(width, _width - width);
 
             if (height != _height)
-            {
                 for (var columnNumber = 0; columnNumber < Columns.Count; columnNumber++)
-                {
                     ResizeColumn(columnNumber, height);
-                }
-            }
 
             _width = width;
             _height = height;
             UpdateCoordinatesList();
         }
 
-        public void SetManyCellsBlocked(IEnumerable<Coordinate2D> coordinates)
+        public void SetManyCellsBlocked(IEnumerable<Coordinate2D> coordinates, bool isBlocked)
         {
             foreach (var coordinate in coordinates)
             {
-                Columns[coordinate.X][coordinate.Y].IsBlocked = true;
+                var cellState = GetCellStateByCoordinate2D(coordinate);
+                cellState.IsBlocked = isBlocked;
             }
 
             UpdateCoordinatesList();
         }
 
-        public void SetOneCellBlocked(Coordinate2D coordinate)
+        public void SetOneCellBlocked(Coordinate2D coordinate, bool isBlocked)
         {
-            SetManyCellsBlocked(new List<Coordinate2D> {coordinate});
+            SetManyCellsBlocked(new List<Coordinate2D> {coordinate}, isBlocked);
         }
 
         public void SetManyCellsMovementType(IEnumerable<Coordinate2D> coordinates, MovementType movementType)
         {
             foreach (var coordinate in coordinates)
             {
-                Columns[coordinate.X][coordinate.Y].MovementType = movementType;
+                var cellState = GetCellStateByCoordinate2D(coordinate);
+                cellState.MovementType = movementType;
             }
         }
 
         public void SetOneCellMovementType(Coordinate2D coordinate, MovementType movementType)
         {
-            SetManyCellsMovementType(new List<Coordinate2D>() {coordinate}, movementType);
+            SetManyCellsMovementType(new List<Coordinate2D> {coordinate}, movementType);
         }
 
         private void UpdateCoordinatesList()
@@ -130,17 +132,17 @@ namespace HexCore.HexGraph
 
         public IEnumerable<Coordinate2D> GetAllCellsOffsetPosition()
         {
-            return Columns.SelectMany(col => col).Select(cell => cell.Position);
+            return Columns.SelectMany(col => col).Select(cell => cell.Coordinate2);
         }
 
         private IEnumerable<Coordinate3D> GetAllCellsCubeCoordinates()
         {
-            return Columns.SelectMany(col => col).Select(cell => cell.Coordinate);
+            return Columns.SelectMany(col => col).Select(cell => cell.Coordinate3);
         }
 
         private IEnumerable<Coordinate3D> GetBlockedCells()
         {
-            return Columns.SelectMany(col => col).Where(cell => cell.IsBlocked).Select(cell => cell.Coordinate);
+            return Columns.SelectMany(col => col).Where(cell => cell.IsBlocked).Select(cell => cell.Coordinate3);
         }
 
         private bool IsInBounds(Coordinate3D coordinate)
@@ -158,45 +160,30 @@ namespace HexCore.HexGraph
             foreach (var direction in _directions)
             {
                 var next = position + direction;
-                if (IsInBounds(next) && !(onlyPassable && IsBlocked(next)))
-                {
-                    yield return next;
-                }
+                if (IsInBounds(next) && !(onlyPassable && IsBlocked(next))) yield return next;
             }
-        }
-
-        public IEnumerable<Coordinate3D> GetPassableNeighbors(Coordinate3D position)
-        {
-            return GetNeighbors(position, true);
-        }
-        
-        public struct Fringe
-        {
-            public Coordinate3D Coordinate;
-            public int CostSoFar;
         }
 
         public List<Coordinate3D> GetMovableArea(Coordinate3D startPosition, int distance, MovementType movementType)
         {
-            // Todo: add movement penalties to this calculation
             var visited = new List<Coordinate3D> {startPosition};
-            var fringes = new List<List<Fringe>> {new List<Fringe> { new Fringe() {Coordinate = startPosition, CostSoFar = 0} }};
+            var fringes = new List<List<Fringe>>
+            {
+                new List<Fringe> {new Fringe {Coordinate = startPosition, CostSoFar = 0}}
+            };
 
             for (var k = 1; k < distance; k++)
             {
                 var fringe = new List<Fringe>();
                 foreach (var position in fringes[k - 1])
+                foreach (var neighbor in GetPassableNeighbors(position.Coordinate))
                 {
-                    foreach (var neighbor in GetPassableNeighbors(position.Coordinate))
-                    {
-                        if (visited.Contains(neighbor)) continue;
-                        var cellState = GetCellStateByCoordinate3D(neighbor);
-                        var movementCostToNeighbor = movementType.GetCostTo(cellState.MovementType.Name);
-                        var totalCost = position.CostSoFar + movementCostToNeighbor;
-                        if (totalCost > distance) continue;
-                        visited.Add(neighbor);
-                        fringe.Add(new Fringe() { Coordinate = neighbor, CostSoFar = totalCost});
-                    }
+                    if (visited.Contains(neighbor)) continue;
+                    var movementCostToNeighbor = GetMovementCost(neighbor, movementType);
+                    var totalCost = position.CostSoFar + movementCostToNeighbor;
+                    if (totalCost > distance) continue;
+                    visited.Add(neighbor);
+                    fringe.Add(new Fringe {Coordinate = neighbor, CostSoFar = totalCost});
                 }
 
                 fringes.Add(fringe);
@@ -207,16 +194,21 @@ namespace HexCore.HexGraph
             return visited;
         }
 
-        private CellState GetCellStateByCoordinate3D(Coordinate3D coordinate)
+        private CellState GetCellStateByCoordinate3D(Coordinate3D coordinate3D)
         {
-            var offset = CoordinateConverter.ConvertOneCubeToOffset(_offsetType, coordinate);
-            return Columns[offset.X][offset.Y];
+            var coordinate2D = CoordinateConverter.ConvertOneCubeToOffset(_offsetType, coordinate3D);
+            return GetCellStateByCoordinate2D(coordinate2D);
         }
 
-        public int GetMovementCost(Coordinate3D coordinate, MovementType unitMovementType)
+        private CellState GetCellStateByCoordinate2D(Coordinate2D coordinate2D)
         {
-            var cellState = GetCellStateByCoordinate3D(coordinate);
-            return unitMovementType.GetCostTo(cellState.MovementType.Name);
+            return Columns[coordinate2D.X][coordinate2D.Y];
+        }
+
+        private struct Fringe
+        {
+            public Coordinate3D Coordinate;
+            public int CostSoFar;
         }
     }
 }
